Trait i2c::Address

pub trait Address: Master {
    fn set_slave_address(
        &mut self, 
        addr: u16, 
        tenbit: bool
    ) -> Result<(), Self::Error>;
}

An I2C master can address different slaves on an I2C bus.

Required Methods

fn set_slave_address(
&mut self,
addr: u16,
tenbit: bool
) -> Result<(), Self::Error>

Sets the current slave to address.

This should not be shifted to include the read/write bit, and therefore should be only 7 bits wide normally.